docker_compose/app/requirements.txt
django==4.0.4
uwsgi==2.0.20
python-dotenv==0.10.5
psycopg2-binary==2.9.4
split_settings==1.0.0
